server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.55.3
python-dateutil==2.8.2
Werkzeug==3.1.3
requests==2.32.3
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
